Fix toString

+++ src/aula08/Pesado.java
@@ -2,9 +2,9 @@ package aula08;
 
 public abstract class Pesado extends Automovel {
     private int numQuadro;
-    private int peso;
+    private double peso;
 
-    public Pesado(String matricula, String marca, String modelo, double potencia, int numQuadro, int peso) {
+    public Pesado(String matricula, String marca, String modelo, double potencia, int numQuadro, double peso) {
         super(matricula, marca, modelo, potencia);
         this.numQuadro = numQuadro;
         this.peso = peso;
@@ -12,7 +12,7 @@ public abstract class Pesado extends Automovel {
 
     @Override
     public String toString() {
-        return String.format("%s; Nº Quadro: %d; Peso: %d",
+        return String.format("%s; Nº Quadro: %d; Peso: %f",
                 super.toString(),
                 this.numQuadro,
                 this.peso);
